+++ R/cgroups.R
@@ -743,26 +743,30 @@ getCGroups1CpuSet <- local({
     expr <- tryCatch({
       parse(text = code)
     }, error = function(ex) {
-      warning(sprintf("Syntax error parsing %s: %s", sQuote(code), sQuote(value0)))
+      warning(sprintf("Syntax error parsing CGroups v1 %s: %s", sQuote(code), sQuote(value0)))
       integer(0L)
     })
   
     value <- tryCatch({
       suppressWarnings(as.integer(eval(expr)))
     }, error = function(ex) {
-      warning(sprintf("Failed to parse %s: %s", sQuote(code), sQuote(value0)))
+      warning(sprintf("Failed to parse CGroups v1 %s: %s", sQuote(code), sQuote(value0)))
       integer(0L)
     })
   
     ## Sanity checks
     max_cores <- maxCores()
-    if (any(value < 0L | value >= max_cores)) {
-      warning(sprintf("[INTERNAL]: Will ignore the cgroups CPU set, because it contains one or more CPU indices that is out of range [0,%d]: %s", max_cores - 1L, value0))
+    if (is.na(max_cores)) {
+      warning(sprintf("[INTERNAL]: Will ignore the CGroups v1 CPU set, because parallel::detectCores() is a missing value: %s", value0))
+      value <- integer(0L)
+      max_cores <- Inf
+    } else if (any(value < 0L | value >= max_cores)) {
+      warning(sprintf("[INTERNAL]: Will ignore the CGroups v1 CPU set, because it contains one or more CPU indices that is out of range [0,%d]: %s", max_cores - 1L, value0))
       value <- integer(0L)
     }
   
     if (any(duplicated(value))) {
-      warning(sprintf("[INTERNAL]: Detected and dropped duplicated CPU indices in the cgroups CPU set: %s", value0))
+      warning(sprintf("[INTERNAL]: Detected and dropped duplicated CPU indices in the CGroups v1 CPU set: %s", value0))
       value <- unique(value)
     }
   
@@ -853,8 +857,11 @@ getCGroups1CpuQuota <- local({
   
     if (!is.na(value)) {
       max_cores <- maxCores()
-      if (!is.finite(value) || value <= 0.0 || value > max_cores) {
-        warning(sprintf("[INTERNAL]: Will ignore the cgroups CPU quota, because it is out of range [1,%d]: %s", max_cores, value))
+      if (is.na(max_cores)) {
+        warning(sprintf("[INTERNAL]: Will ignore the CGroups v1 CPU quota, because parallel::detectCores() is a missing value: %s", value))
+        value <- NA_real_
+      } else if (!is.finite(value) || value <= 0.0 || value > max_cores) {
+        warning(sprintf("[INTERNAL]: Will ignore the CGroups v1 CPU quota, because it is out of range [1,%d]: %s", max_cores, value))
         value <- NA_real_
       }
     }
@@ -915,26 +922,30 @@ getCGroups2CpuSet <- local({
     expr <- tryCatch({
       parse(text = code)
     }, error = function(ex) {
-      warning(sprintf("Syntax error parsing cgroups v2 %s: %s", sQuote(name), sQuote(value0)))
+      warning(sprintf("Syntax error parsing CGroups v2 %s: %s", sQuote(name), sQuote(value0)))
       integer(0L)
     })
 
     value <- tryCatch({
       suppressWarnings(as.integer(eval(expr)))
     }, error = function(ex) {
-      warning(sprintf("Failed to parse cgroups v2 %s: %s", sQuote(name), sQuote(value0)))
+      warning(sprintf("Failed to parse CGroups v2 %s: %s", sQuote(name), sQuote(value0)))
       integer(0L)
     })
 
     ## Sanity checks
     max_cores <- maxCores()
-    if (any(value < 0L | value >= max_cores)) {
-      warning(sprintf("[INTERNAL]: Will ignore the cgroups v2 CPU set, because it contains one or more CPU indices that is out of range [0,%d]: %s", max_cores - 1L, value0))
+    if (is.na(max_cores)) {
+      warning(sprintf("[INTERNAL]: Will ignore the CGroups v2 CPU set, because parallel::detectCores() is a missing value: %s", value0))
+      value <- integer(0L)
+      max_cores <- Inf
+    } else if (any(value < 0L | value >= max_cores)) {
+      warning(sprintf("[INTERNAL]: Will ignore the CGroups v2 CPU set, because it contains one or more CPU indices that is out of range [0,%d]: %s", max_cores - 1L, value0))
       value <- integer(0L)
     }
 
     if (any(duplicated(value))) {
-      warning(sprintf("[INTERNAL]: Detected and dropped duplicated CPU indices in the cgroups v2 CPU set: %s", value0))
+      warning(sprintf("[INTERNAL]: Detected and dropped duplicated CPU indices in the CGroups v2 CPU set: %s", value0))
       value <- unique(value)
     }
 
@@ -1013,8 +1024,11 @@ getCGroups2CpuMax <- local({
     value <- max / period
     if (!is.na(value)) {
       max_cores <- maxCores()
-      if (!is.finite(value) || value <= 0.0 || value > max_cores) {
-        warning(sprintf("[INTERNAL]: Will ignore the cgroups v2 CPU quota, because it is out of range [1,%d]: %s", max_cores, value))
+      if (is.na(max_cores)) {
+        warning(sprintf("[INTERNAL]: Will ignore the CGroups v2 CPU quota, because parallel::detectCores() is a missing value: %s", value))
+        value <- NA_real_
+      } else if (!is.finite(value) || value <= 0.0 || value > max_cores) {
+        warning(sprintf("[INTERNAL]: Will ignore the CGroups v2 CPU quota, because it is out of range [1,%d]: %s", max_cores, value))
         value <- NA_real_
       }
     }